requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
anthropic>=0.25.0
python-dotenv>=1.0.0
aiosqlite>=0.19.0
pydantic>=2.0.0
